requirements.txt
# DEPENDENCIAS DEL PROYECTO - DeclarationLetterOnline
# Instalación: pip install -r requirements.txt

# Framework web
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Base de datos
sqlalchemy==2.0.36
aiosqlite==0.20.0

# Google Gemini AI
google-generativeai>=0.8.0
httpx>=0.27.0

# Groq AI
groq==0.11.0

# Chat con memoria
mem0ai==0.1.28

# Utilidades
python-dotenv==1.0.1
python-dateutil==2.9.0
aiofiles==24.1.0

# Procesamiento de documentos
python-docx==1.1.2

# Validación de datos
pydantic==2.10.3
pydantic-settings==2.6.1

# Manejo de archivos
PyPDF2==3.0.1

# CORS y middleware (starlette is auto-installed by fastapi)
# starlette will be installed automatically with the correct version
